+++ packages/iserver/src/services/data/base/index.ts
@@ -1,0 +1,2 @@
+export * from "./BaseParameter"
+export * from "./FilterParameter"