+++ packages/iserver/src/sm/common/Features.ts
@@ -1,0 +1,8 @@
+import type { Geometry } from "../geometry/Geometry"
+
+export interface Feature {
+  ID: number
+  fieldNames: string[]
+  fieldValues: string[]
+  geometry: Geometry
+}