+++ packages/iserver/src/sm/common/QueryParameter.ts
@@ -1,0 +1,8 @@
+export interface QueryParameter {
+  name: string
+  attributeFilter: string
+  fields?: string[]
+  orderBy?: string
+  groupBy?: string
+  ids?: number[]
+}