+++ packages/iserver/src/sm/common/ServiceResult.ts
@@ -1,0 +1,13 @@
+import type { Feature } from "./Features"
+
+export interface ServiceResult {
+  succeed?: boolean
+  error?: {
+    code: number
+    errorMsg: string
+  }
+  featureCount: number
+  featureUriList: string[]
+  features: Feature[]
+  totalCount: number
+}