+++ packages/iserver/src/sm/geometry/Geometry.ts
@@ -1,25 +1,24 @@
 // type GeometryType = "POINT" | "LINE" | "REGION"
 
+import type { PrjCoordSys } from "../PrjCoordSys"
+import type { Point2D } from "./Point2D"
+
 export enum GeometryType {
   POINT = "POINT",
   LINE = "LINE",
-  REGION = "REGION"
+  REGION = "REGION",
+  REGION3D = "REGION3D"
 }
 
-export interface Geometry<T extends GeometryType> {
-  center: Center
+export interface Geometry<T extends GeometryType = GeometryType> {
+  center: Point2D
   parts: number[]
-  style: null
-  prjCoordSys: null
+  style: any
+  prjCoordSys?: PrjCoordSys | null
   id: number
   type: T
   partTopo: number[]
-  points: Center[]
-}
-
-export interface Center {
-  x: number
-  y: number
+  points: Point2D[]
 }
 
 export interface Point extends Geometry<GeometryType.POINT> {}