+++ packages/iserver/src/sm/geometry/index.ts
@@ -1,0 +1,3 @@
+export * from "./Geometry"
+export * from "./Point2D"
+export * from "./Rectangle2D"