+++ packages/iserver/src/services/data/features/getByBounds.ts
@@ -1,9 +1,9 @@
 import ky from "ky-universal"
 import type { Geometry as GeoJSONGeometry, GeoJsonProperties } from "geojson"
 import type { Options as KyOptions } from "ky-universal"
-import { BaseParameter, FilterParameter, toFeatureResultPayload } from "../base"
-import { filterToQueryParameter } from "../base"
-import type { ServiceResult } from "../../../sm/common/ServiceResult"
+import type { BaseParameter, FilterParameter } from "../base"
+import { filterToQueryParameter , toFeatureResultPayload } from "../base"
+import type { ServiceResult } from "~/sm/common/ServiceResult"
 import { toGeoJSON } from "~/geometry/transformer"
 import type { FeatureResultPayload } from "~/sm/data/featureResults/FeatureResultPayload"
 import { GetFeatureMode } from "~/sm/data/featureResults/GetFeatureMode"
@@ -12,7 +12,6 @@ import type { Rectangle2D } from "~/sm/geometry"
 export interface GetByBoundsParameter extends BaseParameter {
   bounds: Pick<Rectangle2D, "leftBottom" | "rightTop">
   filter?: FilterParameter
-  attributeFilter?: string
 }
 
 export async function getByBounds<G extends GeoJSONGeometry | null = GeoJSONGeometry, P = GeoJsonProperties>(
@@ -21,14 +20,13 @@ export async function getByBounds<G extends GeoJSONGeometry | null = GeoJSONGeom
   kyOptions: KyOptions = {}
 ) {
   const getFeatureMode =
-    options.attributeFilter != null ? GetFeatureMode.BOUNDS_ATTRIBUTEFILTER : GetFeatureMode.BOUNDS
+    options.filter != null ? GetFeatureMode.BOUNDS_ATTRIBUTEFILTER : GetFeatureMode.BOUNDS
 
   const { bounds } = options
 
   const payload: FeatureResultPayload = {
     ...toFeatureResultPayload(options, getFeatureMode),
     ...(options.filter && { queryParameter: filterToQueryParameter(options, options.filter) }),
-    ...(options.attributeFilter && { attributeFilter: options.attributeFilter }),
     bounds: {
       ...bounds,
       bottom: bounds.leftBottom.y,