+++ packages/iserver/src/sm/data/featureResults/FeatureResultPayload.ts
@@ -2,7 +2,7 @@ import type { Rectangle2D } from "../../geometry/Rectangle2D"
 import type { Geometry } from "../../geometry/Geometry"
 import type { SpatialQueryMode } from "../../common/SpatialQueryMode"
 import type { GetFeatureMode } from "./GetFeatureMode"
-import type { PrjCoordSys } from "~/sm/PrjCoordSys"
+import type { PrjCoordSys } from "~/sm/common/PrjCoordSys"
 
 export interface FeatureResultPayload {
   getFeatureMode: GetFeatureMode